WIP working on tests
